modified:   workspaces/DummyRosterTwo/DummyRoster.Common.EntityModel/Models/Product.cs

+++ workspaces/DummyRosterTwo/DummyRoster.Common.EntityModel/Models/Product.cs
@@ -15,7 +15,7 @@ namespace DummyRoster.Common.EntityModel.Models;
 public partial class Product
 {
     [Key]
-    public long Id { get; set; }
+    public int Id { get; set; }
 
     [Column(TypeName = "varchar(32)")]
     public string Name { get; set; } = null!;
@@ -25,10 +25,10 @@ public partial class Product
     public byte[]? Picture { get; set; }
 
     [Column(TypeName = "INT")]
-    public long? CategoryId { get; set; }
+    public int? CategoryId { get; set; }
 
     [Column(TypeName = "INT")]
-    public long? SupplierId { get; set; }
+    public int? SupplierId { get; set; }
 
     [Column(TypeName = "varchar(24)")]
     public string? QuantityPerUnit { get; set; }
@@ -37,13 +37,13 @@ public partial class Product
     public decimal? UnitPrice { get; set; }
 
     [Column(TypeName = "SMALLINT")]
-    public long? UnitsInStock { get; set; }
+    public int? UnitsInStock { get; set; }
 
     [Column(TypeName = "SMALLINT")]
-    public long? UnitsOnOrder { get; set; }
+    public int? UnitsOnOrder { get; set; }
 
     [Column(TypeName = "SMALLINT")]
-    public long? ReorderLevel { get; set; }
+    public int? ReorderLevel { get; set; }
 
     [Column(TypeName = "BOOLEAN")]
     public bool Discontinued { get; set; }
